Access management by model
Allow to indicate in the .config if a model requireUserAccess. Imagine
that you want to only allow some people to access to the server to sync
their data online (like evernote). Well, you say that the _default model
requireUserAccess : ["sync"]. And you update the user's data to give him
access like that : user.access = ["sync"]